Fix thirteenth-round review findings
A broken redirect on every filtered tab
----------------------------------------
redirectFlash appended "?flash=" to whatever path it was given. That is right
for a bare path and silently wrong for one already carrying a query — and the
maintenance due list redirects back to ".../maintenance?scope=overdue". The
result was "?scope=overdue?flash=…": a single parameter whose value is the rest
of the URL. The scope came back as garbage and fell to the default, bouncing
the staffer off the tab they were working, and neither flash nor flash_error was
ever read, so a rejected completion looked exactly like a successful one.

Overdue is the tab the dashboard chip and the nav badge both link to, so this
was the common path — and it had been broken since the feature landed. Every
browser check I ran used the unfiltered scope, which is the one case that works.
The helper now picks its separator; the test drives the real round trip rather
than the path helper alone, and fails against the old code.

Three contained defects
-----------------------
The cost report offered "Settings → Service" to the roles it was built for.
That tab is PermManageSystem; the report is service:view, which finance and
fulfillment hold and admin-only settings are not — so the link 403'd for exactly
the audience the feature names. Gated, with copy that tells them who can.

A garbled maintenance_id was silently discarded: parseOptionalUUID answers nil
for blank and for garbage alike, so a mangled value took the "no maintenance"
branch and the ticket committed unattached, which the comment above it forbids.
Parsed strictly now. And when that attach fails, it reads as a form rejection
instead of a raw JSON 404 with the staffer's typed ticket rolled back and
nothing saying why.

Also
----
store.CloseDueParams still carried TicketID and CloseDue still wrote
`ticket_id = COALESCE(...)` — a dead, unscoped write path for the same
cross-customer hole AttachTicket documents guarding. 81c84cf claimed to have
removed this and removed only the app-level half.

A completion dated in the future was accepted, anchoring the next occurrence
from a day nobody worked. A malformed start date was reported as a missing one.
Parts money was int32 while the labour side had been cast, so the overflow fix
covered one of the two sums. The reschedule moved dates on ended assignments and
retired machines, and counted them in the audit total.

Doc drift: EndAssignment's service comment still said every pending occurrence
goes; 083's Down dropped an index its Up no longer creates; 081's rationale for
omitting an index named an index that could not have served the query anyway;
the round count said eleven and the claimed-but-unmade count said five, when
this commit itself makes them thirteen and eight.

+++ internal/store/service_tickets.go
@@ -619,7 +619,7 @@ func (s *ServiceTicketStore) Totals(ctx context.Context, tx pgx.Tx, ticketID uui
 	var t domain.ServiceTotals
 
 	if err := tx.QueryRow(ctx,
-		`SELECT COALESCE(SUM(unit_cost_cents * quantity), 0) FROM service_parts WHERE ticket_id = $1`,
+		`SELECT COALESCE(SUM(unit_cost_cents::bigint * quantity), 0) FROM service_parts WHERE ticket_id = $1`,
 		ticketID).Scan(&t.PartsCostCents); err != nil {
 		return domain.ServiceTotals{}, fmt.Errorf("sum service parts for ticket %s: %w", ticketID, err)
 	}
@@ -745,7 +745,7 @@ func (s *ServiceTicketStore) CostSummary(ctx context.Context, tx pgx.Tx, f Servi
 	WITH scoped AS (
 	    SELECT id FROM service_tickets WHERE ` + strings.Join(where, " AND ") + `
 	), parts AS (
-	    SELECT p.ticket_id, p.unit_cost_cents * p.quantity AS cents
+	    SELECT p.ticket_id, p.unit_cost_cents::bigint * p.quantity AS cents
 	    FROM service_parts p JOIN scoped s ON s.id = p.ticket_id
 	    WHERE true` + partWindow + `
 	), entries AS (
@@ -847,7 +847,7 @@ func (s *ServiceTicketStore) CostByCustomer(ctx context.Context, tx pgx.Tx, f Se
 
 	query := `
 	WITH combined AS (
-	    SELECT t.customer_id, p.ticket_id, p.unit_cost_cents * p.quantity AS cents,
+	    SELECT t.customer_id, p.ticket_id, p.unit_cost_cents::bigint * p.quantity AS cents,
 	           0 AS minutes, NULL::text AS kind, false AS billable, true AS is_part,
 	           NULL::integer AS rate_cents,
 	           ` + servicePartCostDate + ` AS on_day
